+++ cogs/tracking.py
@@ -564,7 +564,8 @@ class Tracking(commands.Cog, CogUtils):
             await self._watchlist_update_message(self.bot.pool, watchlist, channel, embed)
             await self._watchlist_update_name(watchlist, channel)
         except discord.HTTPException:
-            log.exception(f"{self.tag}[_watchlist_update_content] {watchlist}")
+            # log.exception(f"{self.tag}[_watchlist_update_content] {watchlist}")
+            pass
 
     @staticmethod
     async def _watchlist_update_name(watchlist: Watchlist, channel: discord.TextChannel):
@@ -585,7 +586,7 @@ class Tracking(commands.Cog, CogUtils):
         # We try to get the watched message, if the bot can't find it, we just create a new one
         # This may be because the old message was deleted or this is the first time the list is checked
         try:
-            message = await channel.get_message(watchlist.message_id)
+            message = await channel.fetch_message(watchlist.message_id)
         except discord.HTTPException:
             message = None
         if message is None: